requirements.txt
click==8.0.3
gspread==5.1.1
pre-commit==2.17.0
python-gitlab==3.1.1
